Remove the type conversion in the training

+++ detectinhos/matching.py
@@ -86,15 +86,18 @@ def mine_negatives(
 ) -> torch.Tensor:
     batch_size, num_anchors, _ = positive.shape
     pos_batch, pos_anchor, pos_obj = torch.where(positive)
-    labels = torch.zeros_like(pred[:, :, 0], dtype=torch.long)
+    labels = torch.zeros_like(pred[:, :, 0], dtype=label.dtype)
     labels[pos_batch, pos_anchor] = label[pos_batch, pos_obj].squeeze()
+    # TODO: Avoid type conversion here
     loss = F.cross_entropy(
-        pred.view(-1, pred.shape[-1]), labels.view(-1), reduction="none"
+        pred.view(-1, pred.shape[-1]),
+        labels.long().view(-1),
+        reduction="none",
     ).view(batch_size, num_anchors)
     loss[pos_batch, pos_anchor] = 0
     _, loss_sorted_idx = loss.sort(dim=1, descending=True)
     _, rank = loss_sorted_idx.sort(dim=1)
-    num_pos = positive.sum(dim=(1, 2), dtype=torch.long).unsqueeze(1)
+    num_pos = positive.sum(dim=(1, 2), dtype=label.dtype).unsqueeze(1)
     num_neg = torch.clamp(negpos_ratio * num_pos, max=num_anchors - 1)
     return rank < num_neg.expand_as(rank)
